project file added

+++ src/main/java/com/eao/musiclab/repository/AlbumRepository.java
@@ -3,5 +3,9 @@ package com.eao.musiclab.repository;
 import com.eao.musiclab.model.Album;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AlbumRepository extends JpaRepository<Album, Long> {
+/**
+ * @author kishor
+ * Created on 23/10/2025
+ */
+public interface AlbumRepository extends JpaRepository<Album, Integer> {
 }
