+++ src/main/java/com/eao/musiclab/repository/ArtistRepository.java
@@ -3,5 +3,9 @@ package com.eao.musiclab.repository;
 import com.eao.musiclab.model.Artist;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ArtistRepository extends JpaRepository<Artist, Long> {
+/**
+ * @author kishor
+ * Created on 23/10/2025
+ */
+public interface ArtistRepository extends JpaRepository<Artist, Integer> {
 }
