+++ src/main/java/com/eao/musiclab/repository/GenreRepository.java
@@ -3,5 +3,9 @@ package com.eao.musiclab.repository;
 import com.eao.musiclab.model.Genre;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface GenreRepository extends JpaRepository<Genre, Long> {
+/**
+ * @author kishor
+ * Created on 23/10/2025
+ */
+public interface GenreRepository extends JpaRepository<Genre, Integer> {
 }
