+++ src/main/java/com/eao/musiclab/repository/UserRepository.java
@@ -3,5 +3,9 @@ package com.eao.musiclab.repository;
 import com.eao.musiclab.model.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface UserRepository extends JpaRepository<User, Long> {
+/**
+ * @author kishor
+ * Created on 23/10/2025
+ */
+public interface UserRepository extends JpaRepository<User, Integer> {
 }
